third_party/blink/web_tests/fast/dom/shadow/update-text-of-style-in-shadow-dom-expected.txt

Tests update the text of styles updates styles in shadow dom, crbug.com/247280

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS borderColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
